requirements.txt
einops==0.6.0
numpy==1.23.4
pyntcloud==0.3.1
PyYAML==5.4.1
scikit-learn==1.2.2
pytorch==1.13.1 
pytorch-cuda=11.7
torchvision==0.14.1 
torchaudio==0.13.1 
